+++ src/Resource/SubscribersResource.php
@@ -11,13 +11,16 @@ use Citilink\ExpertSenderApi\Model\SubscribersPostRequest\Options;
 use Citilink\ExpertSenderApi\Model\SubscribersPostRequest\SubscriberInfo;
 use Citilink\ExpertSenderApi\Request\GetSegmentSizeGetRequest;
 use Citilink\ExpertSenderApi\Request\RemovedSubscriberGetRequest;
+use Citilink\ExpertSenderApi\Request\SegmentsGetRequest;
 use Citilink\ExpertSenderApi\Request\SnoozedSubscribersGetRequest;
+use Citilink\ExpertSenderApi\Request\SnoozedSubscribersPostRequest;
 use Citilink\ExpertSenderApi\Request\SubscribersDeleteRequest;
 use Citilink\ExpertSenderApi\Request\SubscribersGetRequest;
 use Citilink\ExpertSenderApi\Request\SubscribersPostRequest;
 use Citilink\ExpertSenderApi\RequestSenderInterface;
 use Citilink\ExpertSenderApi\Response\GetSegmentSizeGetResponse;
 use Citilink\ExpertSenderApi\Response\RemovedSubscribersGetResponse;
+use Citilink\ExpertSenderApi\Response\SegmentsGetResponse;
 use Citilink\ExpertSenderApi\Response\SnoozedSubscribersGetResponse;
 use Citilink\ExpertSenderApi\Response\SubscribersGetEventsHistoryResponse;
 use Citilink\ExpertSenderApi\Response\SubscribersGetFullResponse;
@@ -217,5 +220,47 @@ class SubscribersResource extends AbstractResource
     public function getSubscriberActivity(): SubscriberActivityResource
     {
         return $this->subscriberActivityResource;
+    }
+
+    /**
+     * Get subscriber segments
+     *
+     * List of all subscriber segments defined in the system
+     *
+     * @return SegmentsGetResponse Response
+     */
+    public function getSubscriberSegments(): SegmentsGetResponse
+    {
+        return new SegmentsGetResponse($this->requestSender->send(new SegmentsGetRequest()));
+    }
+
+    /**
+     * Snooze subscriber by Id
+     *
+     * @param int $id Subscriber’s unique identifier
+     * @param int $snoozeWeeks Number of weeks the subscription will be snoozed for (Valid values are 1 to 26)
+     * @param int|null $listId Identifier of list the subscriber will be snoozed on
+     *
+     * @return ResponseInterface Response
+     */
+    public function snoozeSubscriberById(int $id, int $snoozeWeeks, ?int $listId = null): ResponseInterface
+    {
+        return $this->requestSender->send(SnoozedSubscribersPostRequest::createWithId($id, $snoozeWeeks, $listId));
+    }
+
+    /**
+     * Snooze subscriber by Id
+     *
+     * @param string $email Subscriber’s email
+     * @param int $snoozeWeeks Number of weeks the subscription will be snoozed for (Valid values are 1 to 26)
+     * @param int|null $listId Identifier of list the subscriber will be snoozed on
+     *
+     * @return ResponseInterface Response
+     */
+    public function snoozeSubscriberByEmail(string $email, int $snoozeWeeks, ?int $listId = null): ResponseInterface
+    {
+        return $this->requestSender->send(
+            SnoozedSubscribersPostRequest::createWithEmail($email, $snoozeWeeks, $listId)
+        );
     }
 }